В RoomDaoImpl добавил метод поиска свободных комнат в интервале дат по критериям OrderEntity, необходимо добавить комнаты с разными параметрами и покрыть тестами данный метод отель с id=269 как основа для тестов

+++ hotel/trunk/src/main/java/ru/yandex/zhmyd/hotel/model/Order.java
@@ -19,7 +19,7 @@ public class Order {
     private Date endDate;
     private Boolean confirmed;
 
-    private Integer places;
+    private Byte places;
 
     public Long getId() {
         return id;
@@ -89,11 +89,11 @@ public class Order {
         this.roomCategory = roomCategory;
     }
 
-    public Integer getPlaces() {
+    public Byte getPlaces() {
         return places;
     }
 
-    public void setPlaces(Integer places) {
+    public void setPlaces(Byte places) {
         this.places = places;
     }
